feat: theme handling improvements
