+++ src/components/hooks/useSelectedTheme.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { useState, useEffect, useMemo } from 'react'
-import dark from 'providers/theme/theme-dark'
-import light from 'providers/theme/theme-light'
+import dark from 'providers/theme/darkTheme'
+import light from 'providers/theme/lightTheme'
 
 function useTheme(type: 'light' | 'dark' = 'light'): Theme {
   const [theme, setTheme] = useState<Theme>(light)
