+++ src/components/hooks/useTheme.ts
@@ -1,6 +1,6 @@
 import { useState, useEffect } from 'react'
-import dark from 'providers/theme/theme-dark'
-import light from 'providers/theme/theme-light'
+import dark from 'providers/theme/darkTheme'
+import light from 'providers/theme/lightTheme'
 
 function useTheme(type: 'light' | 'dark' = 'light'): Theme {
   const [theme, setTheme] = useState<Theme>(light)
