--- src/providers/theme/theme-dark/colors.ts
@@ -1,21 +1,0 @@
-const colors: Colors = {
-  primary: '#000000',
-  primary1: '#000000',
-  accent: '#E7374C',
-  accent1: '#F18794',
-  text: '#70808C',
-  text1: '#A9B3BA',
-  gray: '#DDDDDD',
-  gray1: '#EBEBEB',
-  gray2: '#F4F4F4',
-  gray3: '#FCFCFC',
-  neutral: '#E3E7EB',
-  neutralLight: '#EEF1F4',
-  neutralBorder: '#F3F5F9',
-  neutralDark: '#C7CED4',
-  neutralTextDisabled: '#ABB8CA',
-  white: '#000000',
-  black: '#FFFFFF',
-}
-
-export default colors
