--- src/providers/theme/theme-dark/index.ts
@@ -1,15 +1,0 @@
-import radius from '../radius'
-import spacing from '../spacing'
-import typography from '../typography'
-import colors from './colors'
-import shadows from './shadows'
-
-const theme: Theme = {
-  colors,
-  radius,
-  shadows,
-  spacing,
-  typography,
-}
-
-export default theme
